+++ examples/symfony-bundles.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    // Other bundles...
+    Tabuna\Map\Symfony\TabunaMapBundle::class => ['all' => true],
+];